meuProjeto/requirements.txt
Flask==2.3.3
google-cloud-firestore==2.11.1
gunicorn==21.2.0